+++ src/apps/vendor-finance/common/vendor-finance.pool-v2.contract-position-fetcher.ts
@@ -1,6 +1,15 @@
 import { Inject } from '@nestjs/common';
+
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
+import {
+  buildDollarDisplayItem,
+  buildPercentageDisplayItem,
+} from '~app-toolkit/helpers/presentation/display-item.present';
+import { getLabelFromToken } from '~app-toolkit/helpers/presentation/image.present';
+import { gqlFetch } from '~app-toolkit/helpers/the-graph.helper';
 import { DefaultDataProps } from '~position/display.interface';
+import { MetaType } from '~position/position.interface';
+import { isBorrowed } from '~position/position.utils';
 import { ContractPositionTemplatePositionFetcher } from '~position/template/contract-position.template.position-fetcher';
 import {
   GetDefinitionsParams,
@@ -10,25 +19,20 @@ import {
   GetTokenBalancesParams,
   GetDataPropsParams,
 } from '~position/template/contract-position.template.types';
+
 import { VendorFinanceContractFactory, VendorFinancePoolV2 } from '../contracts';
+import { IPositionTracker } from '../contracts/ethers/VendorFinancePositionTracker';
+
 import { LENDING_POOLS_V2_QUERY } from './getLendingPoolsQuery';
-import { gqlFetch } from '~app-toolkit/helpers/the-graph.helper';
-import { MetaType } from '~position/position.interface';
-import { getLabelFromToken } from '~app-toolkit/helpers/presentation/image.present';
-import { isBorrowed } from '~position/position.utils';
-import {
-  buildDollarDisplayItem,
-  buildPercentageDisplayItem,
-} from '~app-toolkit/helpers/presentation/display-item.present';
 import {
   VendorFinancePoolDataProps,
   VendorFinancePoolV2Definition,
   VendorLendingPoolsV2GraphResponse,
 } from './vendor-finance.pool.types';
-import { IPositionTracker } from '../contracts/ethers/VendorFinancePositionTracker';
 
 export abstract class VendorFinancePoolV2ContractPositionFetcher extends ContractPositionTemplatePositionFetcher<VendorFinancePoolV2> {
   abstract subgraphUrl: string;
+  abstract positionTrackerAddress: string;
 
   constructor(
     @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
@@ -121,21 +125,26 @@ export abstract class VendorFinancePoolV2ContractPositionFetcher extends Contrac
     ];
   }
 
-  async getDataProps(
-    _params: GetDataPropsParams<VendorFinancePoolV2, DefaultDataProps, VendorFinancePoolV2Definition>,
-  ): Promise<VendorFinancePoolDataProps> {
+  async getDataProps({
+    definition,
+  }: GetDataPropsParams<
+    VendorFinancePoolV2,
+    DefaultDataProps,
+    VendorFinancePoolV2Definition
+  >): Promise<VendorFinancePoolDataProps> {
     return {
-      deployer: _params.definition.deployer,
-      totalDeposited: parseInt(_params.definition.lendBalance) + parseInt(_params.definition.totalBorrowed),
+      deployer: definition.deployer,
+      totalDeposited: parseInt(definition.lendBalance) + parseInt(definition.totalBorrowed),
     };
   }
 
   async getTokenBalancesPerPosition({
     address,
     contractPosition,
+    multicall,
   }: GetTokenBalancesParams<VendorFinancePoolV2, VendorFinancePoolDataProps>) {
-    const collateralToken = contractPosition.tokens[0]!;
-    const lentToken = contractPosition.tokens[1]!;
+    const collateralToken = contractPosition.tokens[0];
+    const lentToken = contractPosition.tokens[1];
     // --- Lender logic ----
     // No deposit, no borrow, but lending out
     if (address === contractPosition.dataProps.deployer.toLowerCase()) {
@@ -144,12 +153,10 @@ export abstract class VendorFinancePoolV2ContractPositionFetcher extends Contrac
     // --! Lender logic !---
 
     // --- Borrower logic ----
-    const positionTrackerAddr = '0x93E73571A71D27cd35a20E14BA5B352C3C2236fC';
     const startKey = '0x0000000000000000000000000000000000000000000000000000000000000000';
-    const multicall = this.appToolkit.getMulticall(this.network);
     const positionTracker = multicall.wrap(
       this.contractFactory.vendorFinancePositionTracker({
-        address: positionTrackerAddr,
+        address: this.positionTrackerAddress,
         network: this.network,
       }),
     );